SLVS-1816 Move ConnectedModeFeaturesConfiguration away from Core assembly and into the ConnectedMode assembly.
This fixes the architecture violation of Core referencing the SonarQube.Client assembly
